+++ packages/xchain-cosmos/src/client.ts
@@ -13,36 +13,39 @@ import {
   TxsPage,
   XChainClient,
   XChainClientParams,
+  singleFee,
 } from '@xchainjs/xchain-client'
-import { Asset, Chain, assetToString, baseAmount } from '@xchainjs/xchain-util'
+import { Asset, BaseAmount, Chain, assetToString, baseAmount, eqAsset } from '@xchainjs/xchain-util'
+import BigNumber from 'bignumber.js'
 
+import { AssetAtom, COSMOS_DECIMAL, DEFAULT_FEE, DEFAULT_GAS_LIMIT } from './const'
 import { CosmosSDKClient } from './cosmos/sdk-client'
 import { TxOfflineParams } from './cosmos/types'
-import { AssetAtom, AssetMuon } from './types'
-import { DECIMAL, getAsset, getDenom, getTxsFromHistory } from './util'
+import { ChainIds, ClientUrls, CosmosClientParams } from './types'
+import {
+  getAsset,
+  getDefaultChainIds,
+  getDefaultClientUrls,
+  getDefaultRootDerivationPaths,
+  getDenom,
+  getTxsFromHistory,
+  protoFee,
+} from './util'
 
 /**
  * Interface for custom Cosmos client
  */
 export interface CosmosClient {
-  getMainAsset(): Asset
   getSDKClient(): CosmosSDKClient
 }
-
-const MAINNET_SDK = new CosmosSDKClient({
-  server: 'https://api.cosmos.network',
-  chainId: 'cosmoshub-4',
-})
-const TESTNET_SDK = new CosmosSDKClient({
-  server: 'https://rest.sentry-02.theta-testnet.polypore.xyz',
-  chainId: 'theta-testnet-001',
-})
 
 /**
  * Custom Cosmos client
  */
 class Client extends BaseXChainClient implements CosmosClient, XChainClient {
-  private sdkClients: Map<Network, CosmosSDKClient> = new Map<Network, CosmosSDKClient>()
+  private sdkClient: CosmosSDKClient
+  private clientUrls: ClientUrls
+  private chainIds: ChainIds
 
   /**
    * Constructor
@@ -57,15 +60,37 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
   constructor({
     network = Network.Testnet,
     phrase,
-    rootDerivationPaths = {
-      [Network.Mainnet]: `44'/118'/0'/0/`,
-      [Network.Testnet]: `44'/118'/0'/0/`,
-      [Network.Stagenet]: `44'/118'/0'/0/`,
-    },
-  }: XChainClientParams) {
+    clientUrls = getDefaultClientUrls(),
+    chainIds = getDefaultChainIds(),
+    rootDerivationPaths = getDefaultRootDerivationPaths(),
+  }: XChainClientParams & CosmosClientParams) {
     super(Chain.Cosmos, { network, rootDerivationPaths, phrase })
-    this.sdkClients.set(Network.Testnet, TESTNET_SDK)
-    this.sdkClients.set(Network.Mainnet, MAINNET_SDK)
+
+    this.clientUrls = clientUrls
+    this.chainIds = chainIds
+
+    this.sdkClient = new CosmosSDKClient({
+      server: this.clientUrls[network],
+      chainId: this.chainIds[network],
+    })
+  }
+
+  /**
+   * Updates current network.
+   *
+   * @param {Network} network
+   * @returns {void}
+   */
+  setNetwork(network: Network): void {
+    // dirty check to avoid using and re-creation of same data
+    if (network === this.network) return
+
+    super.setNetwork(network)
+
+    this.sdkClient = new CosmosSDKClient({
+      server: this.clientUrls[network],
+      chainId: this.chainIds[network],
+    })
   }
 
   /**
@@ -119,7 +144,7 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
   }
 
   getSDKClient(): CosmosSDKClient {
-    return this.sdkClients.get(this.network) || TESTNET_SDK
+    return this.sdkClient
   }
 
   /**
@@ -146,21 +171,6 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
   }
 
   /**
-   * Get the main asset based on the network.
-   *
-   * @returns {string} The main asset based on the network.
-   */
-  getMainAsset(): Asset {
-    switch (this.network) {
-      case Network.Mainnet:
-      case Network.Stagenet:
-        return AssetAtom
-      case Network.Testnet:
-        return AssetMuon
-    }
-  }
-
-  /**
    * Get the balance of a given address.
    *
    * @param {Address} address By default, it will return the balance of the current wallet. (optional)
@@ -168,23 +178,21 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
    * @returns {Balance[]} The balance of the address.
    */
   async getBalance(address: Address, assets?: Asset[]): Promise<Balance[]> {
-    const balances = await this.getSDKClient().getBalance(address)
-    const mainAsset = this.getMainAsset()
+    const coins = await this.getSDKClient().getBalance(address)
+
+    const balances = coins
+      .reduce((acc: Balance[], { denom, amount }) => {
+        const asset = getAsset(denom)
+        return asset ? [...acc, { asset, amount: baseAmount(amount || '0', COSMOS_DECIMAL) }] : acc
+      }, [])
+      .filter(({ asset: balanceAsset }) => !assets || assets.filter((asset) => eqAsset(balanceAsset, asset)).length)
 
     return balances
-      .map((balance) => {
-        return {
-          asset: (balance.denom && getAsset(balance.denom)) || mainAsset,
-          amount: baseAmount(balance.amount, DECIMAL),
-        }
-      })
-      .filter(
-        (balance) => !assets || assets.filter((asset) => assetToString(balance.asset) === assetToString(asset)).length,
-      )
   }
 
   /**
-   * Get transaction history of a given address with pagination options.
+   * Get transaction history of a given address and asset with pagination options.
+   * If `asset` is not set, history will include `ATOM` txs only
    * By default it will return the transaction history of the current wallet.
    *
    * @param {TxHistoryParams} params The options to get transaction history. (optional)
@@ -196,11 +204,12 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
     const limit = (params && params.limit) || undefined
     const txMinHeight = undefined
     const txMaxHeight = undefined
+    const asset = getAsset(params?.asset ?? '') || AssetAtom
+    const messageSender = params?.address ?? this.getAddress()
 
-    const mainAsset = this.getMainAsset()
     const txHistory = await this.getSDKClient().searchTx({
       messageAction,
-      messageSender: (params && params.address) || this.getAddress(),
+      messageSender,
       page,
       limit,
       txMinHeight,
@@ -209,12 +218,12 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
 
     return {
       total: parseInt(txHistory.pagination?.total || '0'),
-      txs: getTxsFromHistory(txHistory.tx_responses || [], mainAsset),
+      txs: getTxsFromHistory(txHistory.tx_responses || [], asset),
     }
   }
 
   /**
-   * Get the transaction details of a given transaction id.
+   * Get the transaction details of a given transaction id. Supports `ATOM` txs only.
    *
    * @param {string} txId The transaction id.
    * @returns {Tx} The transaction details of the given transaction id.
@@ -226,7 +235,7 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
       throw new Error('transaction not found')
     }
 
-    const txs = getTxsFromHistory([txResult], this.getMainAsset())
+    const txs = getTxsFromHistory([txResult], AssetAtom)
     if (txs.length === 0) throw new Error('transaction not found')
 
     return txs[0]
@@ -238,17 +247,32 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
    * @param {TxParams} params The transfer options.
    * @returns {TxHash} The transaction hash.
    */
-  async transfer({ walletIndex, asset, amount, recipient, memo }: TxParams): Promise<TxHash> {
+  async transfer({
+    walletIndex,
+    asset = AssetAtom,
+    amount,
+    recipient,
+    memo,
+    gasLimit = new BigNumber(DEFAULT_GAS_LIMIT),
+    feeAmount = DEFAULT_FEE,
+  }: TxParams & { gasLimit?: BigNumber; feeAmount?: BaseAmount }): Promise<TxHash> {
     const fromAddressIndex = walletIndex || 0
 
-    const mainAsset = this.getMainAsset()
+    const denom = getDenom(asset)
+
+    if (!denom)
+      throw Error(`Invalid asset ${assetToString(asset)} - Only ATOM asset is currently supported to transfer`)
+
+    const fee = protoFee({ denom, amount: feeAmount, gasLimit })
+
     return this.getSDKClient().transfer({
       privkey: this.getPrivateKey(fromAddressIndex),
       from: this.getAddress(fromAddressIndex),
       to: recipient,
-      amount: amount.amount().toString(),
-      asset: getDenom(asset || mainAsset),
+      amount,
+      denom,
       memo,
+      fee,
     })
   }
 
@@ -260,40 +284,56 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
    */
   async transferOffline({
     walletIndex,
-    asset,
+    asset = AssetAtom,
     amount,
     recipient,
     memo,
     from_account_number,
     from_sequence,
+    gasLimit = new BigNumber(DEFAULT_GAS_LIMIT),
+    feeAmount = DEFAULT_FEE,
   }: TxOfflineParams): Promise<string> {
     const fromAddressIndex = walletIndex || 0
 
-    const mainAsset = this.getMainAsset()
+    const denom = getDenom(asset)
+
+    if (!denom)
+      throw Error(`Invalid asset ${assetToString(asset)} - Only ATOM asset is currently supported to transfer`)
+
+    const fee = protoFee({ denom, amount: feeAmount, gasLimit })
+
     return await this.getSDKClient().transferSignedOffline({
       privkey: this.getPrivateKey(fromAddressIndex),
       from: this.getAddress(fromAddressIndex),
       from_account_number,
       from_sequence,
       to: recipient,
-      amount: amount.amount().toString(),
-      asset: getDenom(asset || mainAsset),
+      amount,
+      denom,
       memo,
+      fee,
     })
   }
 
   /**
-   * Get the current fee.
+   * Returns fees.
+   * It tries to get chain fees from THORChain `inbound_addresses` first
+   * If it fails, it returns DEFAULT fees.
    *
-   * @returns {Fees} The current fee.
+   * @returns {Fees} Current fees
    */
   async getFees(): Promise<Fees> {
-    // there is no fixed fee, we set fee amount when creating a transaction.
-    return {
-      type: FeeType.FlatFee,
-      fast: baseAmount(750, DECIMAL),
-      fastest: baseAmount(2500, DECIMAL),
-      average: baseAmount(0, DECIMAL),
+    try {
+      const feeRate = await this.getFeeRateFromThorchain()
+      // convert decimal: 1e8 (THORChain) to 1e6 (COSMOS)
+      // Similar to `fromCosmosToThorchain` in THORNode
+      // @see https://gitlab.com/thorchain/thornode/-/blob/e787022028f662b3a7c594e4a65aca618caa359c/bifrost/pkg/chainclients/gaia/util.go#L86
+      const decimalDiff = COSMOS_DECIMAL - 8 /* THORCHAIN_DECIMAL */
+      const feeRate1e6 = feeRate * 10 ** decimalDiff
+      const fee = baseAmount(feeRate1e6, COSMOS_DECIMAL)
+      return singleFee(FeeType.FlatFee, fee)
+    } catch (error) {
+      return singleFee(FeeType.FlatFee, DEFAULT_FEE)
     }
   }
 }