+++ packages/xchain-cosmos/src/util.ts
@@ -1,14 +1,14 @@
 import { cosmosclient, proto } from '@cosmos-client/core'
-import { FeeType, Fees, Tx, TxFrom, TxTo, TxType } from '@xchainjs/xchain-client'
-import { Asset, assetToString, baseAmount } from '@xchainjs/xchain-util'
+import { cosmos } from '@cosmos-client/core/cjs/proto'
+import { FeeType, Fees, Network, RootDerivationPaths, Tx, TxFrom, TxTo, TxType } from '@xchainjs/xchain-client'
+import { Asset, BaseAmount, CosmosChain, baseAmount, eqAsset } from '@xchainjs/xchain-util'
+import axios from 'axios'
+import BigNumber from 'bignumber.js'
+import Long from 'long'
 
-import { APIQueryParam, RawTxResponse, TxResponse } from './cosmos/types'
-import { AssetAtom, AssetMuon } from './types'
-
-/**
- * The decimal for cosmos chain.
- */
-export const DECIMAL = 6
+import { AssetAtom, COSMOS_DECIMAL, DEFAULT_GAS_LIMIT } from './const'
+import { APIQueryParam, TxResponse, UnsignedTxParams } from './cosmos/types'
+import { ChainId, ChainIds, ClientUrls as ClientUrls } from './types'
 
 /**
  * Type guard for MsgSend
@@ -32,15 +32,14 @@ export const isMsgMultiSend = (msg: unknown): msg is proto.cosmos.bank.v1beta1.M
   (msg as proto.cosmos.bank.v1beta1.MsgMultiSend)?.outputs !== undefined
 
 /**
- * Get denomination from Asset
+ * Get denomination from Asset - currently `ATOM` supported only
  *
  * @param {Asset} asset
  * @returns {string} The denomination of the given asset.
  */
-export const getDenom = (asset: Asset): string => {
-  if (assetToString(asset) === assetToString(AssetAtom)) return 'uatom'
-  if (assetToString(asset) === assetToString(AssetMuon)) return 'umuon'
-  return asset.symbol
+export const getDenom = (asset: Asset): string | null => {
+  if (eqAsset(asset, AssetAtom)) return 'uatom'
+  return null
 }
 
 /**
@@ -51,125 +50,122 @@ export const getDenom = (asset: Asset): string => {
  */
 export const getAsset = (denom: string): Asset | null => {
   if (denom === getDenom(AssetAtom)) return AssetAtom
-  if (denom === getDenom(AssetMuon)) return AssetMuon
+  // IBC assets
+  if (denom.startsWith('ibc/'))
+    // Note: Don't use `assetFromString` here, it will interpret `/` as synth
+    return {
+      chain: CosmosChain,
+      symbol: denom,
+      // TODO (xchain-contributors)
+      // Get readable ticker for IBC assets from denom #600 https://github.com/xchainjs/xchainjs-lib/issues/600
+      // At the meantime ticker will be empty
+      ticker: '',
+      synth: false,
+    }
   return null
 }
 
-const getCoinAmount = (coins?: proto.cosmos.base.v1beta1.ICoin[]) => {
-  return coins
-    ? coins
-        .map((coin) => baseAmount(coin.amount || 0, DECIMAL))
-        .reduce((acc, cur) => baseAmount(acc.amount().plus(cur.amount()), DECIMAL), baseAmount(0, DECIMAL))
-    : baseAmount(0, DECIMAL)
-}
 /**
- * Parse transaction type
+ * Parses amount from `ICoin[]`
+ *
+ * @param {ICoin[]} coinst List of coins
+ *
+ * @returns {BaseAmount} Coin amount
+ */
+const getCoinAmount = (coins: proto.cosmos.base.v1beta1.ICoin[]): BaseAmount =>
+  coins
+    .map((coin) => baseAmount(coin.amount || 0, COSMOS_DECIMAL))
+    .reduce((acc, cur) => baseAmount(acc.amount().plus(cur.amount()), COSMOS_DECIMAL), baseAmount(0, COSMOS_DECIMAL))
+
+/**
+ * Filters `ICoin[]` by given `Asset`
+ *
+ * @param {ICoin[]} coinst List of coins
+ * @param {Asset} asset Asset to filter coins
+ *
+ * @returns {ICoin[]} Filtered list
+ */
+const getCoinsByAsset = (coins: proto.cosmos.base.v1beta1.ICoin[], asset: Asset): proto.cosmos.base.v1beta1.ICoin[] =>
+  coins.filter(({ denom }) => {
+    const coinAsset = !!denom ? getAsset(denom) : null
+    return !!coinAsset ? eqAsset(coinAsset, asset) : false
+  })
+
+/**
+ * Parses transaction history
  *
  * @param {TxResponse[]} txs The transaction response from the node.
- * @param {Asset} mainAsset Current main asset which depends on the network.
- * @returns {Tx[]} The parsed transaction result.
+ * @param {Asset} asset Asset to get history of transactions from
+ *
+ * @returns {Tx[]} List of transactions
  */
-export const getTxsFromHistory = (txs: TxResponse[], mainAsset: Asset): Tx[] => {
-  return txs.reduce((acc, tx) => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    let msgs: Array<proto.cosmos.bank.v1beta1.MsgSend | proto.cosmos.bank.v1beta1.MsgMultiSend>
-    if ((tx.tx as RawTxResponse).body === undefined) {
-      msgs = cosmosclient.codec.packAnyFromCosmosJSON(tx.tx).msg
-    } else {
-      msgs = cosmosclient.codec.packAnyFromCosmosJSON((tx.tx as RawTxResponse).body.messages)
-    }
+export const getTxsFromHistory = (txs: TxResponse[], asset: Asset): Tx[] => {
+  return (
+    txs
+      // order list to have latest txs first in list
+      .sort((a, b) => {
+        if (a.timestamp === b.timestamp) return 0
+        return a.timestamp > b.timestamp ? -1 : 1
+      })
+      .reduce((acc, tx) => {
+        const msgs = tx.tx?.body.messages ?? []
 
-    const from: TxFrom[] = []
-    const to: TxTo[] = []
-    msgs.map((msg) => {
-      if (isMsgSend(msg)) {
-        const msgSend = msg
-        const amount = getCoinAmount(msgSend.amount)
+        const from: TxFrom[] = []
+        const to: TxTo[] = []
+        msgs.map((msg) => {
+          if (isMsgSend(msg)) {
+            const msgSend = msg
+            const coins = getCoinsByAsset(msgSend.amount, asset)
+            const amount = getCoinAmount(coins)
 
-        let from_index = -1
+            let from_index = -1
 
-        from.forEach((value, index) => {
-          if (value.from === msgSend.from_address) from_index = index
-        })
-
-        if (from_index === -1) {
-          from.push({
-            from: msgSend.from_address,
-            amount,
-          })
-        } else {
-          from[from_index].amount = baseAmount(from[from_index].amount.amount().plus(amount.amount()), DECIMAL)
-        }
-
-        let to_index = -1
-
-        to.forEach((value, index) => {
-          if (value.to === msgSend.to_address) to_index = index
-        })
-
-        if (to_index === -1) {
-          to.push({
-            to: msgSend.to_address,
-            amount,
-          })
-        } else {
-          to[to_index].amount = baseAmount(to[to_index].amount.amount().plus(amount.amount()), DECIMAL)
-        }
-      } else if (isMsgMultiSend(msg)) {
-        const msgMultiSend = msg
-
-        msgMultiSend.inputs.map((input) => {
-          const amount = getCoinAmount(input.coins || [])
-
-          let from_index = -1
-
-          from.forEach((value, index) => {
-            if (value.from === input.address) from_index = index
-          })
-
-          if (from_index === -1) {
-            from.push({
-              from: input.address || '',
-              amount,
+            from.forEach((value, index) => {
+              if (value.from === msgSend.from_address) from_index = index
             })
-          } else {
-            from[from_index].amount = baseAmount(from[from_index].amount.amount().plus(amount.amount()), DECIMAL)
+
+            if (from_index === -1) {
+              from.push({
+                from: msgSend.from_address,
+                amount,
+              })
+            } else {
+              from[from_index].amount = baseAmount(
+                from[from_index].amount.amount().plus(amount.amount()),
+                COSMOS_DECIMAL,
+              )
+            }
+
+            let to_index = -1
+
+            to.forEach((value, index) => {
+              if (value.to === msgSend.to_address) to_index = index
+            })
+
+            if (to_index === -1) {
+              to.push({
+                to: msgSend.to_address,
+                amount,
+              })
+            } else {
+              to[to_index].amount = baseAmount(to[to_index].amount.amount().plus(amount.amount()), COSMOS_DECIMAL)
+            }
           }
         })
 
-        msgMultiSend.outputs.map((output) => {
-          const amount = getCoinAmount(output.coins || [])
-
-          let to_index = -1
-
-          to.forEach((value, index) => {
-            if (value.to === output.address) to_index = index
-          })
-
-          if (to_index === -1) {
-            to.push({
-              to: output.address || '',
-              amount,
-            })
-          } else {
-            to[to_index].amount = baseAmount(to[to_index].amount.amount().plus(amount.amount()), DECIMAL)
-          }
-        })
-      }
-    })
-
-    return [
-      ...acc,
-      {
-        asset: mainAsset,
-        from,
-        to,
-        date: new Date(tx.timestamp),
-        type: from.length > 0 || to.length > 0 ? TxType.Transfer : TxType.Unknown,
-        hash: tx.txhash || '',
-      },
-    ]
-  }, [] as Tx[])
+        return [
+          ...acc,
+          {
+            asset,
+            from,
+            to,
+            date: new Date(tx.timestamp),
+            type: from.length > 0 || to.length > 0 ? TxType.Transfer : TxType.Unknown,
+            hash: tx.txhash || '',
+          },
+        ]
+      }, [] as Tx[])
+  )
 }
 
 /**
@@ -193,9 +189,9 @@ export const getQueryString = (params: APIQueryParam): string => {
 export const getDefaultFees = (): Fees => {
   return {
     type: FeeType.FlatFee,
-    fast: baseAmount(750, DECIMAL),
-    fastest: baseAmount(2500, DECIMAL),
-    average: baseAmount(0, DECIMAL),
+    fast: baseAmount(750, COSMOS_DECIMAL),
+    fastest: baseAmount(2500, COSMOS_DECIMAL),
+    average: baseAmount(0, COSMOS_DECIMAL),
   }
 }
 
@@ -206,3 +202,142 @@ export const getDefaultFees = (): Fees => {
  *
  **/
 export const getPrefix = () => 'cosmos'
+
+/**
+ * Default client urls
+ *
+ * @returns {ClientUrls} The client urls for Cosmos.
+ */
+export const getDefaultClientUrls = (): ClientUrls => {
+  const mainClientUrl = 'https://api.cosmos.network'
+  // Note: In case anyone facing into CORS issue, try the following URLs
+  // https://lcd-cosmos.cosmostation.io/
+  // https://lcd-cosmoshub.keplr.app/
+  // @see (Discord #xchainjs) https://discord.com/channels/838986635756044328/988096545926828082/988103739967688724
+  return {
+    [Network.Testnet]: 'https://rest.sentry-02.theta-testnet.polypore.xyz',
+    [Network.Stagenet]: mainClientUrl,
+    [Network.Mainnet]: mainClientUrl,
+  }
+}
+
+/**
+ * Default chain ids
+ *
+ * @returns {ChainIds} Chain ids for Cosmos.
+ */
+export const getDefaultChainIds = (): ChainIds => {
+  const mainChainId = 'cosmoshub-4'
+  return {
+    [Network.Testnet]: 'theta-testnet-001',
+    [Network.Stagenet]: mainChainId,
+    [Network.Mainnet]: mainChainId,
+  }
+}
+
+export const getDefaultRootDerivationPaths = (): RootDerivationPaths => ({
+  [Network.Mainnet]: `44'/118'/0'/0/`,
+  [Network.Testnet]: `44'/118'/0'/0/`,
+  [Network.Stagenet]: `44'/118'/0'/0/`,
+})
+
+export const protoFee = ({
+  denom,
+  amount,
+  gasLimit = new BigNumber(DEFAULT_GAS_LIMIT),
+}: {
+  denom: string
+  amount: BaseAmount
+  gasLimit?: BigNumber
+}): proto.cosmos.tx.v1beta1.Fee =>
+  new proto.cosmos.tx.v1beta1.Fee({
+    amount: [
+      {
+        denom,
+        amount: amount.amount().toFixed(0),
+      },
+    ],
+    gas_limit: Long.fromString(gasLimit.toFixed(0)),
+  })
+
+export const protoMsgSend = ({
+  from,
+  to,
+  amount,
+  denom,
+}: {
+  from: string
+  to: string
+  amount: BaseAmount
+  denom: string
+}): proto.cosmos.bank.v1beta1.MsgSend =>
+  new proto.cosmos.bank.v1beta1.MsgSend({
+    from_address: from,
+    to_address: to,
+    amount: [
+      {
+        amount: amount.amount().toFixed(0),
+        denom,
+      },
+    ],
+  })
+
+export const protoTxBody = ({ from, to, amount, denom, memo }: UnsignedTxParams): proto.cosmos.tx.v1beta1.TxBody => {
+  const msg = protoMsgSend({ from, to, amount, denom })
+
+  return new proto.cosmos.tx.v1beta1.TxBody({
+    messages: [cosmosclient.codec.instanceToProtoAny(msg)],
+    memo,
+  })
+}
+
+export const protoAuthInfo = ({
+  pubKey,
+  sequence,
+  mode,
+  fee,
+}: {
+  pubKey: cosmosclient.PubKey
+  sequence: Long.Long
+  mode: proto.cosmos.tx.signing.v1beta1.SignMode
+  fee?: cosmos.tx.v1beta1.IFee
+}): proto.cosmos.tx.v1beta1.AuthInfo =>
+  new proto.cosmos.tx.v1beta1.AuthInfo({
+    signer_infos: [
+      {
+        public_key: cosmosclient.codec.instanceToProtoAny(pubKey),
+        mode_info: {
+          single: {
+            mode,
+          },
+        },
+        sequence,
+      },
+    ],
+    fee,
+  })
+
+/**
+ * Helper to get Cosmos' chain id
+ * @param {string} url API url
+ */
+export const getChainId = async (url: string): Promise<ChainId> => {
+  const { data } = await axios.get<{ node_info: { network: string } }>(`${url}/node_info`)
+  return data?.node_info?.network || Promise.reject('Could not parse chain id')
+}
+
+/**
+ * Helper to get Cosmos' chain id for all networks
+ * @param {ClientUrl} urls urls (use `getDefaultClientUrl()` if you don't need to use custom urls)
+ */
+export const getChainIds = async (urls: ClientUrls): Promise<ChainIds> => {
+  return Promise.all([
+    getChainId(urls[Network.Testnet]),
+    getChainId(urls[Network.Stagenet]),
+    getChainId(urls[Network.Mainnet]),
+  ]).then(([testnetId, stagenetId, mainnetId]) => ({
+    testnet: testnetId,
+    stagenet: stagenetId,
+    mainnet: mainnetId,
+  }))
+}